some changes

+++ backend/controllers/myPhotosController.js
@@ -1,4 +1,6 @@
-const pool = require('../config/db')
+const { json } = require("express");
+const pool = require("../config/db");
+const jwt = require("jsonwebtoken");
 
 // getting user's photos
 exports.imgsUser = async (req, res) => {
@@ -39,6 +41,13 @@ exports.addImgs = async (req, res) => {
 // delete users images
 exports.deleteImages = async (req, res) => {
   try {
+    const token = req.header("Authorization").split(" ")[1];
+
+    if (!token) {
+      return res.status(403).json({ message: "Token berilmadi" });
+    }
+
+    jwt.verify(token, "MEN SENGA BIR GAP AYTAMAN, HECH KIM BILMASIN");
     const { id } = req.body;
     const result = await pool.query(`DELETE FROM images WHERE id = $1`, [id]);
     res.status(200).json({
@@ -47,6 +56,9 @@ exports.deleteImages = async (req, res) => {
     });
   } catch (error) {
     console.log(error.message);
+    return res.status(403).json({
+      message: error.message
+    });
     res.status(500).send("Girgittonimizda nomaqbul nuqson yuzaga keldi.");
   }
 };
